requirements.txt
python-telegram-bot
aiohttp
python-dotenv
cryptography
